Smart Dustbin
The main objective of the project is to design a smart dustbin which will help in keeping our environment clean and also eco-friendly. We are inspired from Swaach Bharat Mission. Nowadays technologies are getting smarter day-by-day so, as to clean the environment we are designing a smart dustbin by using Arduino. This smart dustbin management system is built on the micro-controller based system having ultrasonic sensors on the dustbin. If dustbin is not maintained than these can cause an unhealthy environment and can cause pollute that affect our health. 
In this proposed technology we have designed a smart dustbin using ARDUINO UNO, along with ultrasonic sensor, servo motor, and battery jumper wire. After all hardware and software connection, now Smart Dustbin program will be run. Dustbin lid will when someone comes near at some range than wait for user to put garbage and close it. It’s properly running or not. For social it will help toward health and hygiene, for business for we try to make it affordable to many as many possible. So that normal people to rich people can take benefit from it.
The Smart Dustbin using Arduino Uno is a project aimed at integrating technology to enhance waste management systems. This report outlines the development and implementation of a smart dustbin system that utilizes Arduino Uno for efficient waste collection and management.